Add files via upload
Project Overview
This is a machine learning project that aims to predict the survival of passengers on the Titanic using a classification model. The dataset provided by Kaggle is used for this project.

Key Features
Exploratory data analysis to understand the dataset and identify relevant features
Preprocessing steps such as handling missing values, encoding categorical features, and scaling numerical features
Implementation of various classification algorithms, including Logistic Regression, Decision Tree, Random Forest, and K-Nearest Neighbors
Hyperparameter tuning to optimize the performance of the models
Evaluation of model performance using appropriate metrics such as accuracy, precision, recall, and F1-score
Comparison of the performance of different classification models
Selection of the best-performing model for the final prediction
How to Use
Clone the repository to your local machine.
Install the required dependencies (e.g., numpy, pandas, scikit-learn).
Open the Jupyter Notebook file and follow the instructions to run the code.
Explore the dataset, preprocess the data, train and evaluate the classification models.
Use the best-performing model to make predictions on the test data.
Results
The best-performing model for this project is the Random Forest Classifier, which achieved an accuracy of approximately 82% on the test data.

Contributions
Contributions to this project are welcome. If you find any issues or have suggestions for improvements, please feel free to create a new issue or submit a pull request.

License
This project is licensed under the MIT License.